dockerfile update --> kleinkram and debugging build_data

+++ box_utils/box_converter/huggingface-dataset-conversion/dataset_builder/build_data.py
@@ -34,6 +34,7 @@ from dataset_builder.utils import messages_in_bag_with_topic
 
 DATA_PREFIX = "data"
 IMAGE_PREFIX = "images"
+DELETE_FOLDER = False
 
 
 ImageExtractorCallback = Callable[
@@ -240,7 +241,9 @@ def _tar_ball_dirs_in_dir(dir: Path) -> None:
             continue
         with tarfile.open(dir / f"{folder}.tar", "w") as tar:
             tar.add(dir / folder, arcname=os.path.basename(folder))
-        shutil.rmtree(dir / folder)
+        # TODO: uncomment this line to remove the original folder
+        if DELETE_FOLDER:
+            shutil.rmtree(dir / folder)
 
 
 def _tar_ball_dataset(base_dataset_path: Path) -> None:
